pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "appd-mcp-server"
version = "1.0.0"
description = "Production-grade MCP server for AppDynamics SaaS Controller"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
    "pydantic>=2.0.0",
    "cachetools>=5.0.0",
    "diskcache>=5.0.0",
    "python-dateutil>=2.9.0",
    "pytz>=2024.1",
    "types-cachetools>=6.2.0.20260408",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest-cov>=5.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["models", "utils", "auth", "client", "parsers", "services", "registries"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
strict = true
python_version = "3.13"
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_any_generics = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short"

[tool.appd_mcp]
api_version = 1

[dependency-groups]
dev = [
    "types-python-dateutil>=2.9.0.20260408",
    "types-pytz>=2026.1.1.20260408",
]
